+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [[Unreleased]]
 
 ### Fixed
-- add `MPTCurrency` support in `Issue` (rippled internal type)
+- Added `MPTCurrency` support in `Issue` (rippled internal type)
 - Fixed the implementation error in get_latest_open_ledger_sequence method. The change uses the "current" ledger for extracting sequence number.
 - Increase default maximum payload size for websocket client
 - Fixed the default behavior of flags field when preparing transactions. By default, flags are not part of the transaction if not explicitly provided.
+- Added support for `amm_info` to `Request.from_dict`
+- Improved error handling for `amm_info`
 
 ### Added
 - Improved validation for models to also check param types
@@ -40,7 +42,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### BREAKING CHANGE:
 
 - Remove Python 3.7 support to fix dependency installation and use 3.8 as new default.
-- Ensure consistent use of ED25519 as the default cryptographic algorithm in `Wallet.from_secret_numbers` method, aligning with changes made in v2.0.0
 
 ### Fixed